ADD keyboard and ports module:
Adding simple keyboard and ports modules.
Ports module provides functions to read and write on a ports using inline asm.
keyboard module provides a function to check if some data are available to read from the keyboard driver, and a keyboard_handler: the function we'll using when the interrupts will be setup.