ADD minimalist keyboard poc:
Adding a really minimalist keyboard management.

The handler functions provide a way to read the keyboard port, and tokenize a scancode taking into account sequences and to send the tokenized scan code to a circular buffer.

The kb_read (Work in progress), will simply call make_break, and then process the char returned.

The make_break function will set up and track the state / lock keys, and return a char if is a valid char.

The map_key function will perform the key mapping from scancode to ascii, according to locks and state keys.