FIX Slab:
Fix move_slab method.
The slabs weren't chained correctly as the link method was not resetting the prev variable to null.
Thus, when unlink was performed, the check of the prev variable to see if the slab was the first on a cache list was not valid and therefore broke the cache list.