UPDATE RTC driver:
Rename RTC driver for CMOS driver as RTC is a part of the CMOS chipset.
Add bin_to_bcd protection as requested by @jgiron42 in PR #186.
Add a RegisterF struct and a function to read it, describing the CMOS's 'Shutdown status' infos.
Add functions to enable update_ended, alarm and periodic RTC interrupt.