FIX slave PIC IRQs wrongly mapped in the IDT:
The IDT entries were not setup correctly for the slave PIC IRQs.
It was due to get_interrupt_id_from_irq in the pic driver which doesn't return the right interrupt id.

This commit also introduces an automatic management of the master IRQ.Slave, when enabling / disabling IRQs of the slave PIC.

fix #187