ADD handle spurious interrupts:
This commit introduces the management of spurious interrupts.

It introduces 2 new variables and their getter to keep track of spurious interrupts in the pic drivers: ``spurious_master: u32`` and ``spurious_slave: u32``.
These variables will be incremented according to whether the interrupt occurred on the master PIC or the slave PIC.

It also introduces a new method ``ack_spurious_interrupt`` to keep ``spurious_master`` and ``spurious_slave`` up-to-date when spurious interrupt occurs, and acknowledge the master if the spurious occurs on the slave.
If this method is called from a real interrupt, it won't do anything and return false.

Since now, the default IRQ handlers will call in any case the ``ack_spurious_interrupt``.

Finally, a shell builtin "spurious" is added to display the amount of spurious interrupts since boot for both master and slave PIC.

fix #189